[SPLTRP-0986]: enhancement(ui): ambient shadow for GroupPocketBalanceCard + extract reusable transition-aware elevation hook (#989)
